requirements.txt
duckdb==1.1.3
pandas==2.2.3
pyarrow==18.1.0
dbt-duckdb==1.8.3
python-dotenv==1.0.1
